Added the ability to GET a sentence for a given show and character
This method is inefficient as all paths are returned from neo4j.
For example, if the sentences "I am a developer" and "I am a
consultant" were provided to the API, a GET would return ["am","am"]
given the word "I". More research to come on ways that this could be
handled by neo4j itself and return only a single word, or even a
full sentence